app/requirements.txt
streamlit==1.37.0
openai
st-pages
python-dotenv
pyyaml
